Agrego Interfaz de carga inicial
Modifique el metodo select action para reciba como parámetro el metodo
de busqueda como un int para usarlo en un case.

+++ TPIA2016/src/frsf/cidisi/exercise/tpia2016/search/actions/IrAula.java
@@ -106,7 +106,7 @@ public class IrAula extends SearchAction {
      */
     @Override
     public Double getCost() {
-        return new Double(0);
+        return new Double(10);
     }
 
     /**
